Moved token management from SpotifyApi to SpotifyApiToken
